+++ src/main/java/de/muspellheim/commons/fx/chart/ScalableAxis.java
@@ -7,10 +7,11 @@ package de.muspellheim.commons.fx.chart;
 
 import javafx.scene.*;
 import javafx.scene.chart.*;
+import javafx.scene.input.*;
 import lombok.*;
 
 /**
- * Zoom a chart axis.
+ * Zoom and move a chart axis.
  */
 public final class ScalableAxis {
 
@@ -20,28 +21,59 @@ public final class ScalableAxis {
     }
 
     /**
-     * Install a scroll handler to zoom in and zoom out a chart axis.
+     * Install scroll handler to zoom in and zoom out a chart axis and drag handler to move a chart axis.
+     * <p>
+     * You can call this function for the x axis and y axis of a xy chart to handle both axis.
+     *
+     * @param chart the chart with the axis
+     * @param axis  the axis to zoom and move
      */
-    public static <X, Y> void install(@NonNull XYChart<X, Y> chart, @NonNull NumberAxis axis) {
-        Cursor cursor = axis.getSide().isHorizontal() ? Cursor.H_RESIZE : Cursor.V_RESIZE;
-        chart.setCursor(cursor);
-        chart.setOnScroll(e -> {
-            double range = axis.getUpperBound() - axis.getLowerBound();
+    public static <T extends Number> void install(@NonNull XYChart<?, ?> chart, @NonNull ValueAxis<T> axis) {
+        installZoomable(chart, axis);
+        installMoveable(chart, axis);
+    }
 
+    private static <T extends Number> void installZoomable(XYChart<?, ?> chart, ValueAxis<T> axis) {
+        chart.addEventHandler(ScrollEvent.SCROLL, e -> {
             Node chartPlotArea = chart.lookup(".chart-plot-background");
-            double relativeX = e.getX() - chartPlotArea.getLayoutX() - chart.getInsets().getLeft();
-            double offset = axis.getValueForDisplay(relativeX).doubleValue();
-            double relativeToOffset = (offset - axis.getLowerBound()) / range;
-
-            double diff = range * 0.1;
-            if (e.getDeltaY() < 0) {
-                diff *= -1;
-            }
-            double newLowerBound = axis.getLowerBound() + diff * relativeToOffset;
-            axis.setLowerBound(newLowerBound);
-            double newUpperBound = axis.getUpperBound() - diff * (1 - relativeToOffset);
-            axis.setUpperBound(newUpperBound);
+            double displayPosition = axis.getSide().isHorizontal()
+                ? e.getX() - chartPlotArea.getLayoutX() - chart.getInsets().getLeft()
+                : e.getY() - chartPlotArea.getLayoutY() - chart.getInsets().getTop();
+            double value = axis.getValueForDisplay(displayPosition).doubleValue();
+            double range = axis.getUpperBound() - axis.getLowerBound();
+            double relativeToOffset = (value - axis.getLowerBound()) / range;
+            double factor = e.getDeltaY() < 0 ? -1 : 1;
+            double delta = range * 0.1 * factor;
+            axis.setLowerBound(axis.getLowerBound() + delta * relativeToOffset);
+            axis.setUpperBound(axis.getUpperBound() - delta * (1 - relativeToOffset));
         });
+    }
+
+    private static <T extends Number> void installMoveable(XYChart<?, ?> chart, ValueAxis<T> axis) {
+        MoveData moveData = new MoveData();
+        chart.addEventHandler(MouseEvent.MOUSE_PRESSED, e -> {
+            moveData.startLowerBound = axis.getLowerBound();
+            moveData.startUpperBound = axis.getUpperBound();
+            double value = axis.getSide().isHorizontal() ? e.getX() : e.getY();
+            moveData.start = axis.getValueForDisplay(value).doubleValue();
+            chart.setCursor(Cursor.CLOSED_HAND);
+        });
+        chart.addEventHandler(MouseEvent.MOUSE_RELEASED, e -> {
+            double value = axis.getSide().isHorizontal() ? e.getX() : e.getY();
+            double end = axis.getValueForDisplay(value).doubleValue();
+            double delta = end - moveData.start;
+            axis.setLowerBound(moveData.startLowerBound - delta);
+            axis.setUpperBound(moveData.startUpperBound - delta);
+            chart.setCursor(null);
+        });
+        // TODO Using mouse dragged instead of mouse release to set bounds is lagging
+        //chart.addEventHandler(MouseEvent.MOUSE_DRAGGED, e -> {});
+    }
+
+    private static class MoveData {
+        private double startLowerBound;
+        private double startUpperBound;
+        private double start;
     }
 
 }